Same configurations on Linux and macOS
Well, docker sucks on macOS, on Linux it runs as it should be. Also, I
miss linux a lot for development, but when I'm using it, I miss macOS as
well, the basic idea here is to have similar environment on both
systems in terms of tooling. Therefore the install script will now run
accordingly with the platform it is initiated.

- General linux configuration for bash and tmux
- Copy and paste via SSH
- Adds prettier command to neovim