Enhance editor settings and update CopilotChat
This commit introduces several enhancements to the Neovim settings.
These include setting linebreak, listchars, spell language, and
correcting typos on common commands. It also disables the clipboard
and sets spell checking for specific filetypes.

In addition, the commit updates the CopilotChat plugin to a newer
commit in the canary branch.

Lastly, it adds German and Portuguese spell files to the nvim/spell
directory.