Add short commands for CopilotChat
This commit introduces short commands for CopilotChat in the Neovim
configuration. It also updates the commit hashes for the telescope.nvim
and ui plugins in the lazy-lock.json file. Additionally, it removes the
theme_toggle option from the chadrc.lua file and corrects the directory
creation command in the mappings.lua file. Lastly, it adds a line break
in the plugins/init.lua file for better readability.